Inline new upsert logic for EquipmentLimitBreak, PawnCraftProgress, StorageItem.
